requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
websockets>=10.0
python-multipart>=0.0.5
pydantic>=1.8.2
numpy>=1.21.0
torch>=1.9.0
transformers>=4.10.0
sounddevice>=0.4.3
scipy>=1.7.0
librosa>=0.8.1
numba>=0.54.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.0
SQLAlchemy>=1.4.23
alembic>=1.7.1
psycopg2-binary>=2.9.1
python-dotenv>=0.19.0
aiofiles>=0.7.0
openai>=0.27.0
soundfile>=0.10.3.post1
webrtcvad>=2.0.10
pyaudio>=0.2.11
wave>=0.0.2